create a _lib for /private-sessions route. move constants & types into that lib.
